requirements.txt
Pillow==10.3.0
protobuf==3.20.3
python-dotenv==1.1.0
streamlit==1.39.0
streamlit_option_menu==0.4.0
google-generativeai==0.8.5